move common.py into a utils folder
